Remove event handler for autofocus on modal shown
This resolves a timing issue affecting the Event Config feature.

Although the Bootstrap documentation advises adding this snippet, the
docs do not account for the fact that we are loading our form via XHR
and inserting it into the DOM. When creating a form dynamically in
Chrome 66, the `autofocus` attribute does apply focus correctly.

What's more, the presence of this handler actually causes a race
condition with Capybara - calling `fill_in` immediately after the
opening of the modal fails more often than not: when the first field to
filled has the `autofocus` attribute then the cursor is prone to being
repositioned in the midst of WebDriver sending keystrokes; when the
first field does not have `autofocus`, it sporadically loses its focus
back to its sibling `autofocus` field.

Similar, but not necessarily related:

- https://github.com/teamcapybara/capybara/issues/1890
- https://github.com/erikras/redux-form/issues/686